requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
charset-normalizer==2.1.0
click==8.1.3
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.9.1
Mako==1.2.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
numpy==1.23.1
openpyxl==3.0.10
pandas==1.4.3
pkgutil_resolve_name==1.3.10
protobuf==3.20.1
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
urllib3==1.26.11
Werkzeug==2.0.2
xlrd==1.2.0
zipp==3.8.1
